move threads to its own file
